requirements.txt
faiss-cpu~=1.10.0
numpy~=1.26.4
sentence-transformers~=3.4.1
torch~=2.6.0
tqdm~=4.67.1